Fix high-frequency reaction path printing
When you try to print the reaction progress too frequently (say every
other xi step), the result is wildly different from what you get for
low-frequency printing.

This was happening because the logic that decides to print jumps over
a bit of logic that sets the scale factor for the next step. If you try
printing every xi step, the scale factor is never updated and the
calculations steps that are just too small.