Fix NaN output in solid solution saturation states
The issue was coming from an overflow. When we calculate maximum entropy
distributions, we (more or less) use the formula

    x_i = \frac{\exp{qk_i}}{\sum_j\exp{qk_i}}

to compute the fraction of each end member in the solid solutions. This
can overflow when the saturation indices get large. To avoid that, we
can instead use the following

    x_i = \frac{1}/(1 + \sum_{j}{\exp{qk_j - qk_i}}),

which doesn't overflow unless the difference between the saturation
indices is large. If that's the case, then it probably would have
overflowed in the original version to.

Whatever the case, this save us from some NaNs!