Use perl instead of sed in tests
On linux, the `sed` utility can do an in-place update with the `-i`
flag. MacOS expects an additional argument (e.g. `-i .bak`) to which to
copy the original file before the in-place update. That means we either
need to check in the test which platform we are on, or just use `perl`.